refactor: ユーザ情報取得処理を別ファイルに分離
